refactor(quality): explicit int->short/char casts, long-tail TUs (#17)
Phase 10 (implicit-int-conversion), long tail — 75 sites across 28 TUs
outside io.c (which lands next as its own commit, the G1/G2 pattern).
Each is a representation-preserving explicit cast of an int (or compound
int expression) into the narrow destination field type (short / schar /
char) — the implicit truncation already happened, the cast only documents
it and silences -Wimplicit-int-conversion. Compound RHS is wrapped whole,
e.g. (schar)(aura * 5). Mirrors ../olympia-g2 73152e4.

Includes the lone -Wimplicit-int-conversion-on-negation site,
scry.c:628 `p->barrier = -(c->who)` -> `(short)(-(c->who))` (a distinct
bracket sub-tag matched by prefix during inventory, so it is fixed here
rather than surfacing at the -Werror lock; G2 used (short)).

This phase is NOT a 64-bit hazard: int->short/char narrowing behaves
identically on ILP32 and LP64. No deliberate golden change — both the
debug and asan-ubsan 206-file manifests stay byte-identical (YES), zero
sanitizer diagnostics.

+++ olympia/add.c
@@ -196,7 +196,7 @@ add_new_player(int pl, char *faction, char *character, char *start_city,
 
   	// by Cappinator: 
     // Changed starting faction noble points
-  	pp->noble_points = 18 + (t/8);
+  	pp->noble_points = (short)(18 + (t/8));
 
 	pp->first_turn = sysclock.turn + 1;
 	pp->last_order_turn = sysclock.turn;
@@ -238,7 +238,7 @@ add_new_player(int pl, char *faction, char *character, char *start_city,
 	// by Cappinator:
 	// Changed starting faction fast study days
 	if (t < 101)
-		p_player(pl)->fast_study = 198 + (t * 2);       /* instant study days */
+		p_player(pl)->fast_study = (short)(198 + (t * 2));       /* instant study days */
 	else
 		p_player(pl)->fast_study = 400;
